+++ app/api/groups/[groupId]/remove-user/route.ts
